Move the Discussion table to a new file
