+++ src/Components/DiscussionTable/DiscussionTable.css
@@ -1,0 +1,19 @@
+.discussion-container {
+  /* Add your specific styles here */
+}
+
+table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+th,
+td {
+  padding: 10px;
+  border: 1px solid #ddd;
+}
+
+th {
+  background-color: #f4f4f4;
+  text-align: left;
+}
