--- src/Pages/Discussions/Discussions-old.css
@@ -1,43 +1,0 @@
-.h1-class {
-    color: white;
-    text-align: center;
-    margin-top: 20px;
-    font-size: 50px;
-}
-
-.ul-class {
-    text-align: center;
-    margin-top: 20px;
-    font-size: 30px;
-}
-
-.background .ul-class .li-class .a-class {
-    text-decoration: none;
-    color: black; 
-}
-
-.li-class {
-    color: red;
-}
-
-.body-class {
-    background: grey;
-}
-
-.background {
-    height: 100%;
-    width: 100%;
-    position: fixed;
-    object-fit: cover;
-    z-index: -1;
-    background-color: rgba(211,45,65, 0.8);
-}
-
-@keyframes fadeIn {
-    from {opacity: 0;}
-    to {opacity: 1;}
-}
-
-.fade-in {
-    animation: fadeIn 500ms ease-in-out;
-}
